movie_management/requirement.txt
argcomplete==1.9.3
boto3==1.11.3
botocore==1.14.3
certifi==2019.11.28
cfn-flip==1.2.2
chardet==3.0.4
Click==7.0
Django==2.2
django-storages==1.8
djangorestframework==3.11.0
docutils==0.15.2
durationpy==0.5
future==0.16.0
hjson==3.0.1
idna==2.8
jmespath==0.9.3
kappa==0.6.0
lambda-packages==0.20.0
placebo==0.9.0
python-dateutil==2.6.1
python-slugify==1.2.4
pytz==2019.3
PyYAML==5.3
requests==2.22.0
s3transfer==0.3.0
six==1.14.0
sqlparse==0.3.0
toml==0.10.0
tqdm==4.19.1
troposphere==2.5.3
Unidecode==1.1.1
urllib3==1.25.7
Werkzeug==0.16.0
wsgi-request-logger==0.4.6
zappa==0.48.2